refactor: passkey service
